Delete FileInput and ImportExport components:
- Removed FileInput component and its index file.
- Removed ImportExport component and its index file.
- Removed unused imports in Routes.tsx and index.ts.
- Relocated exportCSV utility function to a new file in utils/exportCSV.ts.
- Updated Inventory.tsx to import exportCSV from the new location.
