--- frontend/src/components/FileInput/index.ts
@@ -1,1 +1,0 @@
-export * from './FileInput';
