+++ frontend/src/components/Routes/Routes.tsx
@@ -13,7 +13,6 @@ import {
   Organizations,
   SearchPage,
   Settings,
-  TermsOfUse,
   Users,
   Vulnerabilities,
   Vulnerability
