+++ frontend/src/components/index.ts
@@ -2,8 +2,6 @@ export * from './Layout';
 export * from './GovBanner';
 export * from './Header/Header';
 export * from './AuthRoute';
-export * from './FileInput';
-export * from './ImportExport';
 export * from './Routes/RouteGuard';
 export * from './Routes/Routes';
 export * from './FilterDrawer/SearchBar';
